pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "snifferbuddy"
version = "0.1.0"
dependencies = [
    "paho-mqtt",
    "python-dotenv",
    "anyio",
    "PyYAML",
    "pydantic",
    "duckdb"
]

[tool.setuptools]
package-dir = {"" = "."}
packages = {find = {where = [".", "src"], exclude = ["tests*"]}}

[tool.pytest.ini_options]
addopts = "-v"
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
